requirement.txt
streamlit
google-genai
pillow
pandas
